Added and Implemented quality_score in Project entity

+++ src/main/java/com/shubham/codeforgeai/model/CodeFile.java
@@ -24,11 +24,11 @@ public class CodeFile {
 
     private int lineCount;
 
-    private long methodCount;
+    private int methodCount;
 
-    private long classCount;
+    private int classCount;
 
-    private long complexityScore;
+    private int complexityScore;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "project_id", nullable = false)
